requirements.txt
# EOG Cursor Control - Python Dependencies
# Install: pip install -r requirements.txt

# Serial communication
pyserial>=3.5

# Mouse/keyboard control
pyautogui>=0.9.54

# Signal processing
scipy>=1.10.0
numpy>=1.24.0

# Machine learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Data handling
pandas>=2.0.0

# Visualization (optional, for debugging)
matplotlib>=3.7.0

# Keyboard input for data collection
pynput>=1.7.6
